+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- unified API version resolution for extra operations (promote, reject, reset, revert, file, status) to use dynamic version selection
 
 ## [5.8.0] - 2026-03-31